+++ src/cmd_read.cpp
@@ -95,7 +95,7 @@ static bool readLine(ARState &state, int fd, const ArrayObject &argvObj, unsigne
       continue;
     }
     if (unlikely(strBuf.size() == StringObject::MAX_SIZE)) {
-      raiseError(state, TYPE::OutOfRangeError, ERROR_STRING_LIMIT);
+      raiseStringLimit(state);
       return false;
     }
     strBuf += static_cast<char>(ch);