+++ src/object.cpp
@@ -505,7 +505,7 @@ bool Value::appendAsStr(ARState &state, StringRef value) {
   const bool small = isSmallStr(this->kind());
   const size_t size = small ? smallStrSize(this->kind()) : typeAs<StringObject>(*this).size();
   if (unlikely(size > StringObject::MAX_SIZE - value.size())) {
-    raiseError(state, TYPE::OutOfRangeError, ERROR_STRING_LIMIT);
+    raiseStringLimit(state);
     return false;
   }