+++ CHANGELOG.md
@@ -85,6 +85,7 @@
     - when call ``ARState_readLine``, always be foreground process
 - fix ``Module#_fullname`` method when pass valid fully qualified command name
 - fix expansion error check
+- add missing string size checking to ``CLI#usage`` method
 
 ## [0.32.1] - 2024-01-21